+++ modules/dxp/apps/akismet/akismet-service/src/main/java/com/liferay/akismet/internal/client/constants/AkismetConstants.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.akismet.client.constants;
+package com.liferay.akismet.internal.client.constants;
 
 /**
  * @author Jamie Sammonss
