+++ modules/dxp/apps/akismet/akismet-service/src/main/java/com/liferay/akismet/internal/service/AkismetMBMessageLocalServiceWrapper.java
@@ -15,8 +15,8 @@
 package com.liferay.akismet.internal.service;
 
 import com.liferay.akismet.client.AkismetClient;
-import com.liferay.akismet.client.constants.AkismetConstants;
 import com.liferay.akismet.configuration.AkismetServiceConfiguration;
+import com.liferay.akismet.internal.client.constants.AkismetConstants;
 import com.liferay.akismet.model.AkismetEntry;
 import com.liferay.akismet.service.AkismetEntryLocalService;
 import com.liferay.message.boards.constants.MBMessageConstants;
